Implemented a lot of things around tags

+++ src/objects/TagsTab.gd
@@ -2,13 +2,16 @@ extends Control
 @export var tag_viewer : Control
 @onready var main_node = get_tree().get_root().get_node("Main")
 @export var edit_button : Button
+@export var delete_button : Button
 
 func _ready():
-	pass # Replace with function body.
+	_on_tag_viewer_tag_pressed_sig()
 
 
-func refrash_items_list(tag_data):
-	tag_viewer.refrash_items_list(tag_data)
+func refrash_tags_list(tag_data):
+	tag_viewer.refrash_tags_list(tag_data)
+	await get_tree().process_frame
+	_on_tag_viewer_tag_pressed_sig()
 
 
 func _on_create_tag_button_pressed():
@@ -20,11 +23,22 @@ func _on_test_selected_tags_button_pressed():
 	#print(tag_viewer.get_selected_id())
 
 
+func confirme_action(what_to_do):
+	if what_to_do == Global.WhatToDo.Delete:
+		var selected_tags = tag_viewer.get_selected_ids()
+		if selected_tags.size() != 0:
+			main_node.delete_tags(selected_tags)
+	await get_tree().process_frame
+	_on_tag_viewer_tag_pressed_sig()
+	
+
 func _on_delete_selected_tags_button_pressed():
-	var selected_tags = tag_viewer.get_selected_ids()
-	if selected_tags.size() != 0:
-		#TODO warning message with number of items with tags to delete
-		main_node.delete_tags(selected_tags)
+	var number_of_items = \
+		main_node.get_number_of_items_with_tags(tag_viewer.get_selected_ids())
+	var mess = ""
+	if number_of_items > 0:
+		mess = str(number_of_items) +  " items use tags that will be removed"
+	main_node.confirme_action_dialogue(self, Global.WhatToDo.Delete, mess)
 
 
 func _on_edit_tags_button_pressed():
@@ -34,4 +48,6 @@ func _on_edit_tags_button_pressed():
 
 
 func _on_tag_viewer_tag_pressed_sig():
-	edit_button.disabled = tag_viewer.get_selected_ids().size() != 1
+	var selected_ids = tag_viewer.get_selected_ids().size()
+	edit_button.disabled = selected_ids != 1
+	delete_button.disabled = selected_ids == 0
